Removed cleanPartitionsOlderThan from Snapshot source
