+++ core-plugins/src/main/java/co/cask/hydrator/plugin/batch/source/SnapshotFileBatchAvroSource.java
@@ -100,7 +100,7 @@ public class SnapshotFileBatchAvroSource extends SnapshotFileBatchSource<AvroKey
     private String schema;
 
     public SnapshotAvroConfig(String name, @Nullable String basePath, String schema) {
-      super(name, basePath, null, null);
+      super(name, basePath, null);
       this.schema = schema;
     }
   }
