+++ core-plugins/src/main/java/co/cask/hydrator/plugin/batch/source/SnapshotFileBatchParquetSource.java
@@ -99,7 +99,7 @@ public class SnapshotFileBatchParquetSource extends SnapshotFileBatchSource<Null
     private String schema;
 
     public SnapshotParquetConfig(String name, @Nullable String basePath, String schema) {
-      super(name, basePath, null, null);
+      super(name, basePath, null);
       this.schema = schema;
     }
   }
